scripts/requirements.txt
kafka-python==1.4.4
psycopg2-binary==2.7.6.1
